refactor(adr): move extractAdrFields into adr-format.ts so the module is wired in
The dead-code ratchet (scripts/check-dead-code.sh, #1540) fails a src/ module that
nothing in src/ imports. adr-format.ts had only a test file, because its consumer
(the templateFormat rewiring) is deferred to #1647.

Fix by reversing the shared-helper direction instead of adding a placeholder caller:
extractAdrFields + ExtractedAdrFields now live in adr-format.ts (the natural home for
ADR-format concerns), and adr-discovery.ts imports them. So the heavily-used discovery
module depends on adr-format.ts (dead count holds at 5, no rise), with no cycle
(adr-format imports nothing from discovery) and no change to #1647's files.

typecheck green; adr-format + adr-discovery-tags tests green (23); dead-code OK.

Refs #1466, #1540.
